Zillion: move client to worlds/zillion (#2649)
